create new components for the lists
